removing public PE endpoint fallback
